+++ src/main/kotlin/blackjack/domain/blackjack/BlackJackGame.kt
@@ -1,4 +1,9 @@
-package blackjack.domain
+package blackjack.domain.blackjack
+
+import blackjack.domain.card.CardDeck
+import blackjack.domain.participants.Dealer
+import blackjack.domain.participants.Guest
+import blackjack.domain.participants.User
 
 class BlackJackGame {
     lateinit var getCommand: (String) -> String
