+++ src/main/kotlin/blackjack/domain/card/Card.kt
@@ -1,3 +1,3 @@
-package blackjack.domain
+package blackjack.domain.card
 
 data class Card(val mark: CardMark, val value: CardValue)
