+++ src/main/kotlin/blackjack/domain/card/CardMark.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 enum class CardMark {
     CLOVER,
