+++ src/main/kotlin/blackjack/domain/card/CardValue.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 enum class CardValue(val value: Int) {
     ACE(1),
