+++ src/main/kotlin/blackjack/domain/participants/Participants.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.participants
 
 data class Participants(
     val dealer: Dealer,
