+++ src/test/kotlin/blackjack/domain/blackjack/BlackJackBuilderTest.kt
@@ -1,5 +1,6 @@
-package blackjack.domain
+package blackjack.domain.blackjack
 
+import blackjack.domain.card.Cards
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
