+++ src/test/kotlin/blackjack/domain/card/CardDeckTest.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
